Remove submodule to use live version of pyre-check
Previously we used a very old state of pyre-check repository as a
submodule to run Github Actions against and test the action in this
repository.

Currently, pyre-check has moved forward and current python module of
pyre-check fails to detect issues based on the submodule. Also, since we
want the action to be upto date with pyre-check it only makes sense to
use git clone during the workflows instead of a submodule fixed on a
particular instance of the pyre-check repository.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>